Paginate the data from `getMedia`
Without changing the API, the data from `getMedia` will get paged.

When `getMedia` is called, it stores the album identifier. If the
identifier is the same than the identifier stored at previous call, the
next page of data is returned.

Each page contains 20 assets. The `creationDate` of the oldest asset is
stored at every call and this is used to get older assets at the next call.